/************* GHMC-Hardware v.55 *************/
			       27-08-2017 

* Updated Scene logic for gathering the specific controllers
* Added Program Change Switch And Program Change Toggle
* ToDo logic and fade in/out calculation

+++ src/controller/switch/ControlChangeFadeSwitch16Bit.h
@@ -8,7 +8,7 @@
 #ifndef SRC_CONTROLLER_SWITCH_CONTROLCHANGEFADESWITCH16BIT_H_
 #define SRC_CONTROLLER_SWITCH_CONTROLCHANGEFADESWITCH16BIT_H_
 
-#include "../controller.h"
+#include "../common/controller.h"
 
 class ControlChangeFadeSwitch16Bit: public Controller {
 public:
@@ -43,7 +43,7 @@ private:
     uint16_t fadeIn;
     uint16_t fadeOut;
 
-    uint16_t parameter;
+    uint8_t parameter;
 
     Dispatcher* dispatcher;
 };
